Added a lot of files. Moved a lot of things around
Moved my working directory into git, so now hopefully the latest from badger
and the latest from coyote are together and working.

I probably broke a lot of things, especially the build scripts for hymn_verse
they will have to be cleaned up eventually

I added the book files, as the html is kind of hard to get otherwise
it takes a lot of time to download it.  The scripts will need to be updated
in there and eventually we could use a refresh of the data