Updated content from the web site. Got new books.
The format of the website changed quite a bit, and in particular they have added
chapters properly now! So the chapter logic and display changed a lot.
I rewrote the chapter logic, but we would have to re-run the entire set of books
if we want to put everything in the new format.  I didn't do this... That means
that some books are in the old orig HTML format, and some are in the new format.
The list of books can be found in books/README.